Adjust subtitle font size based on text length
Modified the `subtitle` font size to be conditional based on the length of the `text2` string. If `text2` has 3 or more characters, the font size is set to 30f; otherwise, it remains at 35f. The vertical position for drawing `text2` has also been updated to reflect this change, using 29 or 22 as appropriate.